requirements.txt
backcall==0.2.0
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
decorator==4.4.2
docopt==0.6.2
h5py==3.1.0
idna==3.4
imageio==2.15.0
importlib-resources==5.4.0
ipdb==0.13.13
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.3.1
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
olefile==0.46
opencv-python==4.5.5.62
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
prompt-toolkit==3.0.36
protobuf==3.19.4
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
scikit-image==0.17.2
scipy==1.5.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboardX==2.5.1
tifffile==2020.9.3
tomli==1.2.3
torch==1.10.1
torchaudio==0.10.1
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.11.2
tornado==6.1
tqdm==4.63.0
traitlets==4.3.3
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.15
visdom==0.2.4
wcwidth==0.2.6
websocket-client==1.3.1
yarg==0.1.9
zipp==3.6.0
